--- src/components/Layout/NotificationBell.jsx
@@ -1,1 +1,0 @@
-export { default } from "../notifications/NotificationBell";
